requirements.txt
dnspython==2.7.0
Flask==2.2.5
ipwhois==1.3.0
joblib==1.2.0
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
Requests==2.32.3
scikit_learn==1.2.2
whois==1.20240129.2
